print redirect location
this is an attempt at solving dkhamsing/awesome_bot#30 that could
use a little feedback before making a PR

note that all of the headers from the faraday request are saved.
to print the redirect location, only the location header is used,
but all are saved to make any headers that are needed in the future
easier to find

also note that the travis test case actually doesn't test this
since it needs the --allow-redirect flag (travis will fail without it)
but the err msg isn't actually printed unless that flag is left off

also note that if statement to determine a redirect status is
repeated from the Result class. this violates DRY and should be
cleaned up.