Replace regex for parsing relative links
This requires the part in parentheses to directly follow the
part in square brackets.

The link parsing logic is still quite complex, and there are
edge cases that still won't work, so it may be worth replacing
it with a markdown parser instead. For now I've just tried
to make it handle a few more cases correctly.

For more details of valid and invalid links in markdown,
see https://spec.commonmark.org/0.28/#links