Remove test for broken Strict-Transport-Security header
This test no longer works because the URL no longer returns that header:

    curl -I 'http://okeowoaderemi.com/site/assets/files/1103/zf2-flowchart.jpg'
    HTTP/1.1 200 OK
    Date: Sat, 14 Jul 2018 11:33:13 GMT
    Server: Apache
    X-Frame-Options: SAMEORIGIN
    Last-Modified: Sat, 31 Jan 2015 21:38:32 GMT
    Accept-Ranges: bytes
    Content-Length: 1437460
    X-XSS-Protection: 1; mode=block
    Content-Type: image/jpeg

The HTTPS version of the website doesn't return it either (and also has a broken certificate).

Since this test was testing a particular website's broken HTTP response, and other
websites are unlikely to be broken in the same way, I've just removed it.

The header is invalid for two reasons:

- Fancy quote marks are not valid syntax for this header
- This header should not be set on HTTP websites anyway (browsers will ignore it)

See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Strict-Transport-Security